LU-3069 build: fix 'integer handling' issues
Fix 'program hangs' defects found by Coverity version 6.5.1:
Improper use of negative value (NEGATIVE_RETURNS)
Variable is passed to a parameter that cannot be negative.
Macro compares unsigned to 0 (NO_EFFECT)
This comparison is always true.
Operands don't affect result (CONSTANT_EXPRESSION_RESULT)
Test is always false regardless of the values of its operands.
Wrong operator used (CONSTANT_EXPRESSION_RESULT)
Test is always true regardless of the values of its operand.

Reviewed-on: http://review.whamcloud.com/5888
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/mdt/mdt_handler.c
@@ -4843,7 +4843,7 @@ static int mdt_init0(const struct lu_env *env, struct mdt_device *m,
         const char                *identity_upcall = "NONE";
         struct md_device          *next;
         int                        rc;
-        int                        node_id;
+	long                       node_id;
         mntopt_t                   mntopts;
         ENTRY;
